FW: reduce the loop counters to 32-bit
We can never loop more than 2^31 times anyway, because of this check in
`max_loop_count_exceeded()`:
```c++
    // unsigned comparisons so sApp->current_max_loop_count == -1 causes an always false
    if (unsigned(data->inner_loop_count) >= unsigned(sApp->shmem->current_max_loop_count))
        return true;
```

In any case, if the test is calling `test_{time,loop}_condition()` more
than 2^32 times, it's doing something wrong. Our test design target is
that a single test loop is between 1 to 100 ms, which would allow a test
to run for 49.7 days before overflowing.

+++ framework/test_data.h
@@ -74,8 +74,8 @@ struct alignas(64) Main : Common
 struct alignas(64) TestCommon : Common
 {
     /* Number of iterations of the inner loop (aka #times test_time_condition called) */
-    uint64_t inner_loop_count;
-    uint64_t inner_loop_count_at_fail;
+    uint32_t inner_loop_count;
+    uint32_t inner_loop_count_at_fail;
 
     /* Thread ID */
     std::atomic<tid_t> tid;
